compartmentalize statuscard
